OPENNLP-1876: Address review round 2: rename to SocialMediaCharSequenceNormalizer, adversarial URL pins, spellcheck helper cleanup
TwitterCharSequenceNormalizer is renamed to
SocialMediaCharSequenceNormalizer; the old name stays as a deprecated
alias so existing references keep working, and the langdetect factory
chain uses the new name. The URL normalizer gains a differential test
running adversarial URL shapes (userinfo, ports, percent escapes, IPv6,
backslashes, punycode, stray schemes) through the former patterns and
the scan side by side. The spellcheck helpers become instance methods
with self-contained javadoc, the package-private test seam documents
why it is package private, and the editorial guard comment is dropped.
